requirements.txt
Flask==1.1.2
numpy==1.19.5
Pillow==2.2.2